[move] gather extension
Util 패키지의 extension을 하나로 수정
