+++ app/src/main/java/org/android/go/sopt/util/extension/ViewExt.kt
@@ -1,7 +1,8 @@
-package org.android.go.sopt.util
+package org.android.go.sopt.util.extension
 
 import android.graphics.Color
 import android.view.View
+import android.widget.Toast
 import com.google.android.material.snackbar.Snackbar
 
 fun View.makeSnackBar(message: String, isShort: Boolean = true) {
@@ -12,4 +13,7 @@ fun View.makeSnackBar(message: String, isShort: Boolean = true) {
         .show()
 }
 
-
+fun View.makeToast(message: String, isShort: Boolean = true) {
+    val duration = if (isShort) Toast.LENGTH_SHORT else Toast.LENGTH_LONG
+    Toast.makeText(context, message, duration).show()
+}
